Python config (and much more)

+++ lobster/commands/plot.py
@@ -70,6 +70,9 @@ def unix2matplotlib(time):
 def unpack(arg):
     source, target = arg
     try:
+        if os.path.isfile(target):
+            logger.info("skipping {0}".format(source))
+            return
         logger.info("unpacking {0}".format(source))
         with open(target, 'w') as output:
             input = gzip.open(source, 'rb')
@@ -77,8 +80,6 @@ def unpack(arg):
             input.close()
     except IOError:
         logger.error("cannot unpack {0}".format(source))
-        return False
-    return True
 
 def mp_call(arg):
     fct, args, kwargs = arg
@@ -240,16 +241,15 @@ class Plotter(object):
     PROF = 8
 
     def __init__(self, config, outdir=None):
-        self.__workdir = config['workdir']
+        self.config = config
         self.__store = unit.UnitStore(config)
 
-        util.verify(self.__workdir)
-        self.__id = config['id']
+        util.verify(self.config.workdir)
 
         if outdir:
             self.__plotdir = outdir
         else:
-            self.__plotdir = config.get("plotdir", self.__id)
+            self.__plotdir = config.plotdir if config.plotdir else self.config.label
         self.__plotdir = os.path.expandvars(os.path.expanduser(self.__plotdir))
 
         if not os.path.isdir(self.__plotdir):
@@ -279,7 +279,7 @@ class Plotter(object):
 
     def readdb(self):
         logger.debug('reading database')
-        db = sqlite3.connect(os.path.join(self.__workdir, 'lobster.db'), timeout=90)
+        db = sqlite3.connect(os.path.join(self.config.workdir, 'lobster.db'), timeout=90)
         stats = {}
 
         wflow_ids = {}
@@ -404,34 +404,7 @@ class Plotter(object):
                     ('workdir_footprint', 'i4')
                     ])
 
-        summary_data = list(db.execute("""
-                select
-                    label,
-                    events,
-                    (select sum(events_read) from tasks where status in (2, 6, 8) and type = 0 and workflow = workflows.id),
-                    (select sum(events_written) from tasks where status in (2, 6, 8) and type = 0 and workflow = workflows.id),
-                    units + units_masked,
-                    units,
-                    units_done,
-                    units_paused,
-                    '' || round(
-                            units_done * 100.0 / units,
-                        1) || ' %'
-                from workflows"""))
-        summary_data += list(db.execute("""
-                select
-                    'Total',
-                    sum(events),
-                    (select sum(events_read) from tasks where status in (2, 6, 8) and type = 0),
-                    (select sum(events_written) from tasks where status in (2, 6, 8) and type = 0),
-                    sum(units + units_masked),
-                    sum(units),
-                    sum(units_done),
-                    sum(units_paused),
-                    '' || round(
-                            sum(units_done) * 100.0 / sum(units),
-                        1) || ' %'
-                from workflows"""))
+        summary_data = list(self.__store.workflow_status())[1:]
 
         # for cases where units per task changes during run, get per-unit info
         total_units = 0
@@ -469,7 +442,7 @@ class Plotter(object):
         if filename:
             fn = filename
         else:
-            fn = os.path.join(self.__workdir, 'lobster_stats.log')
+            fn = os.path.join(self.config.workdir, 'lobster_stats.log')
 
         with open(fn) as f:
             headers = dict(map(lambda (a, b): (b, a), enumerate(f.readline()[1:].split())))
@@ -517,31 +490,23 @@ class Plotter(object):
 
         res = {}
         for label in processed:
-            jsondir = os.path.join('jsons', label)
             if not os.path.exists(os.path.join(self.__plotdir, jsondir)):
                 os.makedirs(os.path.join(self.__plotdir, jsondir))
             lumis = LumiList(lumis=processed[label])
-            lumis.writeJSON(os.path.join(self.__plotdir, jsondir, 'processed.json'))
-            res[label] = [(os.path.join(jsondir, 'processed.json'), 'processed')]
+            lumis.writeJSON(os.path.join(jsondir, 'processed_{}.json'.format(label)))
+            res[label] = [(os.path.join('jsons', 'processed_{}.json'.format(label)), 'processed')]
 
-            published = os.path.join(self.__workdir, label, 'published.json')
+            published = os.path.join(self.config.workdir, label, 'published.json')
             if os.path.isfile(published):
-                shutil.copy(published, os.path.join(self.__plotdir, jsondir))
-                res[label] += [(os.path.join(jsondir, 'published.json'), 'published')]
+                shutil.copy(published, os.path.join(jsondir, 'published_{}.json'.format(label)))
+                res[label] += [(os.path.join('jsons', 'published_{}.json'.format(label)), 'published')]
 
         return res
 
-    def savelogs(self, failed_tasks, samples=5, subdir=''):
+    def savelogs(self, failed_tasks, samples=5):
+        logdir = os.path.join(self.__plotdir, 'logs')
         work = []
         codes = {}
-
-        logdir = os.path.join(self.__plotdir, subdir, 'logs')
-        if os.path.exists(logdir):
-            for dirpath, dirnames, filenames in os.walk(logdir):
-                logs = [os.path.join(dirpath, fn) for fn in filenames if fn.endswith('.log')]
-                map(os.unlink, logs)
-        else:
-            os.makedirs(logdir)
 
         for exit_code, tasks in zip(*split_by_column(failed_tasks[['id', 'exit_code']], 'exit_code')):
             if exit_code == 0:
@@ -554,7 +519,7 @@ class Plotter(object):
                 codes[exit_code][1][id] = []
 
                 try:
-                    source = glob.glob(os.path.join(self.__workdir, '*', 'failed', util.id2dir(id)))[0]
+                    source = glob.glob(os.path.join(self.config.workdir, '*', 'failed', util.id2dir(id)))[0]
                 except IndexError:
                     continue
 
@@ -570,7 +535,7 @@ class Plotter(object):
                         codes[exit_code][1][id].append(l[:-3])
                         work.append([s, t])
 
-        for label, _, _, _, _, _, _, paused, _ in self.__store.workflow_status()[1:]:
+        for label, _, _, _, _, _, _, paused, failed, skipped, _ in list(self.__store.workflow_status())[1:-1]:
             if paused == 0:
                 continue
 
@@ -578,8 +543,8 @@ class Plotter(object):
             skipped = self.__store.skipped_files(label)
 
             for id in failed:
-                source = os.path.join(self.__workdir, label, 'failed', util.id2dir(id))
-                target = os.path.join(logdir, label, 'failed')
+                source = os.path.join(self.config.workdir, label, 'failed', util.id2dir(id))
+                target = os.path.join(logdir, 'failed_' + label)
                 if not os.path.exists(target):
                     os.makedirs(target)
 
@@ -591,7 +556,7 @@ class Plotter(object):
                         work.append([s, t])
 
             if len(skipped) > 0:
-                outname = os.path.join(logdir, label, 'skipped_files.txt')
+                outname = os.path.join(logdir, 'skipped_{}.txt'.format(label))
                 if not os.path.isdir(os.path.dirname(outname)):
                     os.makedirs(os.path.dirname(outname))
                 with open(outname, 'w') as f:
@@ -865,7 +830,7 @@ class Plotter(object):
 
 
         if len(failed_tasks) > 0:
-            logs = self.savelogs(failed_tasks, subdir=subdir)
+            logs = self.savelogs(failed_tasks)
 
             fail_labels, fail_values = split_by_column(failed_tasks, 'exit_code', threshold=0.025)
 
@@ -1038,12 +1003,14 @@ class Plotter(object):
         # Templating
         # ----------
 
+        categories = [c.name for c in self.config.categories if c.name != 'merge']
+
         env = jinja2.Environment(loader=jinja2.FileSystemLoader(
             os.path.join(os.path.dirname(__file__), 'data')))
         env.filters["datetime"] = lambda d: datetime.fromtimestamp(d).strftime('%a, %d %b %Y, %H:%M')
         env.tests["sum"] = lambda s: s == "Total"
         overview = env.get_template('index.html')
-        wflow = env.get_template('workflow.html')
+        wflow = env.get_template('category.html')
 
         jsons = self.savejsons(processed_lumis)
 
@@ -1052,7 +1019,7 @@ class Plotter(object):
 
         with open(os.path.join(self.__plotdir, 'index.html'), 'w') as f:
             f.write(overview.render(
-                id=self.__id,
+                id=self.config.label,
                 plot_time=time.time(),
                 plot_starttime=self.__xmin,
                 plot_endtime=self.__xmax,
@@ -1063,19 +1030,27 @@ class Plotter(object):
                 bad_tasks=len(failed_tasks) > 0,
                 good_tasks=len(success_tasks) > 0,
                 foremen=foremen_names,
-                workflows=sorted(wflow_ids.keys())
+                categories=categories
             ).encode('utf-8'))
 
         # -----------------------
-        # Workflow specific plots
+        # Category specific plots
         # -----------------------
+        logdir = os.path.join(self.__plotdir, 'logs')
+        if os.path.exists(logdir):
+            for dirpath, dirnames, filenames in os.walk(logdir):
+                logs = [os.path.join(dirpath, fn) for fn in filenames if fn.endswith('.log')]
+                map(os.unlink, logs)
+        else:
+            os.makedirs(logdir)
+
         outdir = os.path.join(self.__plotdir, 'all')
         if not os.path.exists(outdir):
             os.makedirs(outdir)
         logs = self.make_workflow_plots('all', edges, good_tasks, failed_tasks, success_tasks, merge_tasks, xmin, xmax)
         with open(os.path.join(self.__plotdir, 'all', 'index.html'), 'w') as f:
             f.write(wflow.render(
-                id=self.__id,
+                id=self.config.label,
                 label='all workflows',
                 bad_tasks=len(failed_tasks) > 0,
                 good_tasks=len(success_tasks) > 0,
@@ -1084,18 +1059,28 @@ class Plotter(object):
                 jsons=jsons,
                 bad_logs=logs,
                 foremen=foremen_names,
-                workflows=sorted(wflow_ids.keys())
+                categories=categories
             ).encode('utf-8'))
 
-        for label, id_ in wflow_ids.items():
+        for category in self.config.categories:
+            label = category.name
+            if label == 'merge':
+                continue
+            ids = []
+            labels = []
+            for workflow in self.config.workflows:
+                if workflow.category == category:
+                    ids.append(wflow_ids[workflow.label])
+                    labels.append(workflow.label)
+
             outdir = os.path.join(self.__plotdir, label)
             if not os.path.exists(outdir):
                 os.makedirs(outdir)
 
-            wf_good_tasks = good_tasks[good_tasks['workflow'] == id_]
-            wf_failed_tasks = failed_tasks[failed_tasks['workflow'] == id_]
-            wf_success_tasks = success_tasks[success_tasks['workflow'] == id_]
-            wf_merge_tasks = merge_tasks[merge_tasks['workflow'] == id_]
+            wf_good_tasks = good_tasks[np.in1d(good_tasks['workflow'], ids)]
+            wf_failed_tasks = failed_tasks[np.in1d(failed_tasks['workflow'], ids)]
+            wf_success_tasks = success_tasks[np.in1d(success_tasks['workflow'], ids)]
+            wf_merge_tasks = merge_tasks[np.in1d(merge_tasks['workflow'], ids)]
 
             logs = self.make_workflow_plots(label, edges,
                     wf_good_tasks,
@@ -1106,16 +1091,16 @@ class Plotter(object):
 
             with open(os.path.join(self.__plotdir, label, 'index.html'), 'w') as f:
                 f.write(wflow.render(
-                    id=self.__id,
+                    id=self.config.label,
                     label=label,
                     bad_tasks=len(wf_failed_tasks) > 0,
                     good_tasks=len(wf_success_tasks) > 0,
                     merge_tasks=len(wf_merge_tasks) > 0,
-                    summary=[xs for xs in summary_data if xs[0] == label],
+                    summary=[xs for xs in summary_data if xs[0] in labels],
                     jsons=jsons,
                     bad_logs=logs,
                     foremen=foremen_names,
-                    workflows=sorted(wflow_ids.keys())
+                    categories=categories
                 ).encode('utf-8'))
 
         p = multiprocessing.Pool(10)
